+++ CHANGELOG.md
@@ -1,9 +1,19 @@
 # Change Log
 
-All notable changes to the "find-unused-files" extension will be documented in this file.
+## [v0.0.4]
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+- Add Result Panel instead of log file
+- Add rerun button
+- Add progress bar to show progress
 
-## [Unreleased]
+## [v0.0.3]
+
+- Fix minor bugs
+
+## [v0.0.2]
+
+- Fix minor bugs
+
+## [v0.0.1]
 
 - Initial release